backend/requirements.txt
fastapi==0.110.3
uvicorn==0.29.0
motor==3.5.1
pydantic==1.10.15
python-dotenv==1.0.1